requirements.txt
edgartools>=4.23.0
pandas
pyyaml
requests
openai>=1.0.0
